+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG
 ---------
 
 #### Next Release
+- Added missing on click listener for simple reading mode
 
 #### v0.4.0 (2019-05-22)
 - Added "complex" reading mode to show bookmark & note with verse